+++ app/src/main/java/com/pocketshell/app/composer/PromptComposerSheet.kt
@@ -53,7 +53,6 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.IconButtonDefaults
-import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.ModalBottomSheet
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.SheetState
@@ -120,8 +119,6 @@ import com.pocketshell.app.voice.PendingTranscriptionItem
 import com.pocketshell.app.voice.UndeliveredTranscript
 import dagger.hilt.internal.GeneratedComponentManagerHolder
 import com.pocketshell.core.agents.AgentKind
-import com.pocketshell.core.storage.entity.PendingTranscriptionEntity
-import com.pocketshell.uikit.components.DisclosureIcon
 import com.pocketshell.uikit.theme.LocalPocketShellSemantic
 import com.pocketshell.uikit.theme.PocketShellColors
 import com.pocketshell.uikit.theme.PocketShellTheme
@@ -1217,36 +1214,21 @@ internal fun SheetContent(
                 Spacer(modifier = Modifier.height(8.dp))
             }
 
-            // Issue #180: pending-transcription banner + expandable list.
-            // Renders only when at least one row is queued; sits above the
-            // mic row so the user notices it before re-recording.
-            if (pendingItems.isNotEmpty()) {
-                PendingTranscriptionsBanner(
-                    items = pendingItems,
-                    retryingIds = state.retryingIds,
-                    expanded = pendingListExpanded,
-                    onToggle = onTogglePendingList,
-                    onRetry = onRetryPending,
-                    onDiscard = onDiscardPending,
-                    onSaveAsAudio = onSavePendingAsAudio,
-                )
-                Spacer(modifier = Modifier.height(8.dp))
-            }
-
-            // Issue #900: queued outbound prompts must be visible before send
-            // is wired to enqueue. This read/delete surface is intentionally
-            // foreground-only and scoped to the current composer target.
-            if (outboundQueueItems.isNotEmpty()) {
-                OutboundQueueBanner(
-                    items = outboundQueueItems,
-                    expanded = outboundQueueExpanded,
-                    onToggle = onToggleOutboundQueue,
-                    onDelete = onDeleteOutboundItem,
-                    onRetry = onRetryOutboundItem,
-                    onResendAll = onResendAllOutbound,
-                )
-                Spacer(modifier = Modifier.height(8.dp))
-            }
+            PromptComposerQueueBanners(
+                pendingItems = pendingItems,
+                retryingIds = state.retryingIds,
+                pendingListExpanded = pendingListExpanded,
+                onTogglePendingList = onTogglePendingList,
+                onRetryPending = onRetryPending,
+                onDiscardPending = onDiscardPending,
+                onSavePendingAsAudio = onSavePendingAsAudio,
+                outboundQueueItems = outboundQueueItems,
+                outboundQueueExpanded = outboundQueueExpanded,
+                onToggleOutboundQueue = onToggleOutboundQueue,
+                onDeleteOutboundItem = onDeleteOutboundItem,
+                onRetryOutboundItem = onRetryOutboundItem,
+                onResendAllOutbound = onResendAllOutbound,
+            )
 
             // Issue #566: staged attachments render as compact square tiles at
             // the bottom of the composer (above the controls row), ChatGPT /
@@ -1873,16 +1855,6 @@ private fun DiscardRecordingButton(
 // intentional token rather than off-ladder drift. See docs/design-system.md.
 private val ComposerActionPillRadius = 22.dp
 private val ComposerActionPillShape = RoundedCornerShape(ComposerActionPillRadius)
-
-// Genuine sub-ladder "micro role" geometry: the outbound-queue action buttons
-// (the per-row Resend affordance and the #1308 "Resend all" banner) use a
-// 6dp radius (design-system.md "Radius, Elevation, And Borders" — "Micro role
-// badges | 3-6dp"), smaller than the named ladder rungs (8/14/20/28dp). 6dp
-// matches the sibling per-row queue button; named here (vs an inline literal)
-// so it stays one intentional token rather than off-ladder drift. See
-// docs/design-system.md.
-private val ComposerQueueButtonRadius = 6.dp
-private val ComposerQueueButtonShape = RoundedCornerShape(ComposerQueueButtonRadius)
 
 // Issue #1152: one baseline HEIGHT for EVERY recording-row pill (Discard /
 // Insert / Send) so the row reads as a single deliberate control ladder instead
@@ -2620,528 +2592,6 @@ internal fun barEnvelopeHeightDp(index: Int): Float {
     return 6f + envelope * 22f // [6dp .. 28dp]
 }
 
-/**
- * Issue #900: committed outbound prompts for the current session. This is a
- * foreground visibility/action surface only; durable retry behavior lands in
- * the owning VM/store slice.
- */
-@Composable
-private fun OutboundQueueBanner(
-    items: List<OutboundItem>,
-    expanded: Boolean,
-    onToggle: () -> Unit,
-    onDelete: (String) -> Unit,
-    onRetry: (String) -> Unit,
-    onResendAll: () -> Unit,
-) {
-    val collapsedRetryItem = if (expanded) null else retryableOutboundQueueItem(items)
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .background(
-                color = PocketShellColors.SurfaceElev,
-                shape = RoundedCornerShape(8.dp),
-            )
-            .border(
-                width = 1.dp,
-                color = PocketShellColors.BorderSoft,
-                shape = RoundedCornerShape(8.dp),
-            )
-            .testTag(COMPOSER_OUTBOUND_QUEUE_BANNER_TAG),
-    ) {
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable(role = androidx.compose.ui.semantics.Role.Button, onClick = onToggle)
-                .padding(horizontal = 12.dp, vertical = 10.dp)
-                .testTag(COMPOSER_OUTBOUND_QUEUE_TOGGLE_TAG),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = outboundQueueHeadline(items),
-                    color = PocketShellColors.Text,
-                    fontSize = 13.sp,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                Text(
-                    text = outboundQueueSubline(items),
-                    color = PocketShellColors.TextSecondary,
-                    fontSize = 11.sp,
-                    maxLines = 1,
-                    overflow = TextOverflow.Ellipsis,
-                )
-            }
-            if (collapsedRetryItem != null) {
-                Spacer(modifier = Modifier.width(8.dp))
-                PendingActionButton(
-                    label = "Retry",
-                    primary = true,
-                    onClick = { onRetry(collapsedRetryItem.id) },
-                    modifier = Modifier.testTag(
-                        composerOutboundQueueRetryTestTag(collapsedRetryItem.id),
-                    ),
-                )
-            }
-            Spacer(modifier = Modifier.width(8.dp))
-            DisclosureIcon(
-                expanded = expanded,
-                tint = PocketShellColors.TextSecondary,
-            )
-        }
-
-        if (expanded) {
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 12.dp)
-                    .height(1.dp)
-                    .background(PocketShellColors.BorderSoft),
-            )
-            // Issue #1308: batch "Resend all" — re-arm EVERY resendable row
-            // (Queued/Failed) at once, in FIFO order, instead of tapping each
-            // row's Retry. Shown ONLY when at least TWO rows are actually
-            // resendable: a single resendable row already has its own tap-to-retry,
-            // so a batch affordance there would be the #971/#987 double-affordance
-            // confusion. Rows still delivering (InFlight/Uploading) are not counted.
-            val resendableCount = items.count {
-                it.state == OutboundState.Queued || it.state == OutboundState.Failed
-            }
-            if (resendableCount >= 2) {
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 12.dp, vertical = 8.dp)
-                        .clip(ComposerQueueButtonShape)
-                        .background(PocketShellColors.Accent, ComposerQueueButtonShape)
-                        .clickable(
-                            role = androidx.compose.ui.semantics.Role.Button,
-                            onClick = onResendAll,
-                        )
-                        .padding(vertical = 10.dp)
-                        .testTag(COMPOSER_OUTBOUND_QUEUE_RESEND_ALL_TAG),
-                    contentAlignment = Alignment.Center,
-                ) {
-                    Text(
-                        text = "Resend all ($resendableCount)",
-                        color = PocketShellColors.OnAccent,
-                        fontSize = 13.sp,
-                        fontWeight = FontWeight.SemiBold,
-                    )
-                }
-                Box(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(horizontal = 12.dp)
-                        .height(1.dp)
-                        .background(PocketShellColors.BorderSoft),
-                )
-            }
-            items.forEach { item ->
-                OutboundQueueRow(
-                    item = item,
-                    onDelete = { onDelete(item.id) },
-                    onRetry = { onRetry(item.id) },
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun OutboundQueueRow(
-    item: OutboundItem,
-    onDelete: () -> Unit,
-    onRetry: () -> Unit,
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 12.dp, vertical = 8.dp)
-            .testTag(composerOutboundQueueItemRowTestTag(item.id)),
-    ) {
-        Text(
-            text = formatRelativeTimestamp(item.createdAtMs, System.currentTimeMillis()),
-            color = PocketShellColors.TextSecondary,
-            style = MaterialTheme.typography.labelSmall,
-            fontWeight = FontWeight.SemiBold,
-        )
-        Spacer(modifier = Modifier.height(2.dp))
-        Text(
-            text = outboundQueueStateLabel(item),
-            color = PocketShellColors.TextSecondary,
-            fontSize = 11.sp,
-        )
-        if (item.cleanText.isNotBlank()) {
-            Spacer(modifier = Modifier.height(4.dp))
-            Text(
-                text = item.cleanText,
-                color = PocketShellColors.Text,
-                style = PocketShellType.bodyDense,
-                maxLines = 2,
-                overflow = TextOverflow.Ellipsis,
-            )
-        }
-        if (item.attachments.isNotEmpty()) {
-            Spacer(modifier = Modifier.height(2.dp))
-            Text(
-                text = outboundAttachmentCountLabel(item.attachments.size),
-                color = PocketShellColors.TextSecondary,
-                fontSize = 11.sp,
-            )
-        }
-        if (item.state == OutboundState.Queued || item.state == OutboundState.Failed) {
-            Spacer(modifier = Modifier.height(8.dp))
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.spacedBy(8.dp, Alignment.End),
-            ) {
-                PendingActionButton(
-                    label = "Delete",
-                    primary = false,
-                    onClick = onDelete,
-                    modifier = Modifier.testTag(composerOutboundQueueDeleteTestTag(item.id)),
-                )
-                PendingActionButton(
-                    label = "Retry",
-                    primary = true,
-                    onClick = onRetry,
-                    modifier = Modifier.testTag(composerOutboundQueueRetryTestTag(item.id)),
-                )
-            }
-        }
-    }
-}
-
-/**
- * Issue #180: banner + expandable list rendered above the mic row when
- * the failed/offline-queued transcription list is non-empty.
- *
- * Two visual states:
- *  - **Collapsed (default):** a single row showing "N pending
- *    transcription(s)" with a chevron. Tapping anywhere on the row
- *    toggles to the expanded state. The row also surfaces the first
- *    queued row's status ("Waiting for network" vs the Whisper error
- *    text) so the user can decide whether to expand before tapping.
- *  - **Expanded:** the collapsed row plus a vertical list of per-item
- *    rows. Each per-item row has:
- *      - The relative timestamp ("X minutes ago")
- *      - The status / error message
- *      - Retry / Discard buttons when below the retry cap
- *      - Save-as-audio / Discard buttons when at the cap
- *
- * The whole banner is wrapped in a single tap target for the
- * collapsed/expand toggle, and individual buttons stop propagation by
- * using `clickable(...)` directly on themselves.
- */
-@Composable
-private fun PendingTranscriptionsBanner(
-    items: List<PendingTranscriptionItem>,
-    retryingIds: Set<String>,
-    expanded: Boolean,
-    onToggle: () -> Unit,
-    onRetry: (String) -> Unit,
-    onDiscard: (String) -> Unit,
-    onSaveAsAudio: (String) -> Unit,
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .background(
-                color = PocketShellColors.AccentSoft,
-                shape = RoundedCornerShape(8.dp),
-            )
-            .border(
-                width = 1.dp,
-                color = PocketShellColors.Accent.copy(alpha = 0.4f),
-                shape = RoundedCornerShape(8.dp),
-            )
-            .testTag(COMPOSER_PENDING_BANNER_TAG),
-    ) {
-        // Summary row — always rendered, always tappable.
-        Row(
-            modifier = Modifier
-                .fillMaxWidth()
-                .clickable(role = androidx.compose.ui.semantics.Role.Button, onClick = onToggle)
-                .padding(horizontal = 12.dp, vertical = 10.dp)
-                .testTag(COMPOSER_PENDING_TOGGLE_TAG),
-            verticalAlignment = Alignment.CenterVertically,
-        ) {
-            Column(modifier = Modifier.weight(1f)) {
-                Text(
-                    text = pendingSummaryHeadline(items),
-                    color = PocketShellColors.Accent,
-                    fontSize = 13.sp,
-                    fontWeight = FontWeight.SemiBold,
-                )
-                val subline = pendingSummarySubline(items)
-                if (subline.isNotEmpty()) {
-                    Spacer(modifier = Modifier.height(2.dp))
-                    Text(
-                        text = subline,
-                        color = PocketShellColors.Accent.copy(alpha = 0.85f),
-                        fontSize = 11.sp,
-                    )
-                }
-            }
-            DisclosureIcon(
-                expanded = expanded,
-                tint = PocketShellColors.Accent,
-            )
-        }
-
-        if (expanded) {
-            // 1dp top divider so the per-item list reads as a sub-section
-            // rather than running into the summary row.
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(horizontal = 12.dp)
-                    .height(1.dp)
-                    .background(PocketShellColors.Accent.copy(alpha = 0.25f)),
-            )
-            items.forEach { item ->
-                PendingTranscriptionRow(
-                    item = item,
-                    retrying = item.id in retryingIds,
-                    onRetry = { onRetry(item.id) },
-                    onDiscard = { onDiscard(item.id) },
-                    onSaveAsAudio = { onSaveAsAudio(item.id) },
-                )
-            }
-        }
-    }
-}
-
-/**
- * Issue #180: per-item row inside the expanded queue banner.
- */
-@Composable
-private fun PendingTranscriptionRow(
-    item: PendingTranscriptionItem,
-    retrying: Boolean,
-    onRetry: () -> Unit,
-    onDiscard: () -> Unit,
-    onSaveAsAudio: () -> Unit,
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(horizontal = 12.dp, vertical = 8.dp)
-            .testTag(composerPendingItemRowTestTag(item.id)),
-    ) {
-        Text(
-            text = formatRelativeTimestamp(item.recordingTimestampMs, System.currentTimeMillis()),
-            color = PocketShellColors.Accent,
-            fontSize = 12.sp,
-            fontWeight = FontWeight.SemiBold,
-        )
-        // Issue #688: a retry in flight shows immediate "Retrying…" feedback
-        // so the Retry tap is never a silent no-op. The status line takes
-        // priority over the prior error/waiting text while the round-trip
-        // is open.
-        val statusText = when {
-            retrying -> PENDING_RETRYING_MESSAGE
-            item.isWaitingForNetwork -> PendingTranscriptionItem.NETWORK_WAITING_MESSAGE
-            item.lastErrorMessage != null -> item.lastErrorMessage
-            else -> "Queued — tap retry to send"
-        }
-        Spacer(modifier = Modifier.height(2.dp))
-        Text(
-            text = statusText,
-            color = PocketShellColors.Accent.copy(alpha = 0.85f),
-            fontSize = 11.sp,
-        )
-        if (item.retryCount > 0 && !item.atRetryCap) {
-            Spacer(modifier = Modifier.height(2.dp))
-            Text(
-                text = "Attempt ${item.retryCount + 1} of " +
-                    "${PendingTranscriptionEntity.MAX_RETRY_ATTEMPTS}",
-                color = PocketShellColors.Accent.copy(alpha = 0.7f),
-                fontSize = 11.sp,
-            )
-        }
-        Spacer(modifier = Modifier.height(8.dp))
-        // #208 follow-up move #2: right-align the action row so the
-        // primary button (Retry or Save audio) lands closest to the
-        // right thumb. Discard (secondary) renders to its left.
-        // Maintainer feedback 2026-05-27 + audit comment-4554970633
-        // call this out — Retry was previously leftmost under the
-        // default Arrangement.Start.
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.spacedBy(8.dp, Alignment.End),
-        ) {
-            if (!item.atRetryCap) {
-                // Order matters: with Arrangement.End the row fills
-                // right-to-left visually but children still render in
-                // declaration order. Discard first → ends up left of
-                // Retry; Retry last → anchors the right edge.
-                PendingActionButton(
-                    label = "Discard",
-                    primary = false,
-                    onClick = onDiscard,
-                    modifier = Modifier.testTag(composerPendingDiscardTestTag(item.id)),
-                )
-                PendingActionButton(
-                    // Issue #688: the button label flips to "Retrying…" and
-                    // the click is gated while a round-trip is in flight, so
-                    // a second tap cannot stack a duplicate retry on top.
-                    label = if (retrying) "Retrying…" else "Retry",
-                    primary = true,
-                    enabled = !retrying,
-                    onClick = onRetry,
-                    modifier = Modifier.testTag(composerPendingRetryTestTag(item.id)),
-                )
-            } else {
-                // 3-retry cap hit — Whisper has repeatedly failed for
-                // this audio. Offer Save (so the user can hand it to a
-                // different transcription tool) or Discard. Save is the
-                // primary action here and travels to the right edge.
-                PendingActionButton(
-                    label = "Discard",
-                    primary = false,
-                    onClick = onDiscard,
-                    modifier = Modifier.testTag(composerPendingDiscardTestTag(item.id)),
-                )
-                PendingActionButton(
-                    label = "Save audio",
-                    primary = true,
-                    onClick = onSaveAsAudio,
-                    modifier = Modifier.testTag(composerPendingSaveTestTag(item.id)),
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun PendingActionButton(
-    label: String,
-    primary: Boolean,
-    onClick: () -> Unit,
-    modifier: Modifier = Modifier,
-    enabled: Boolean = true,
-) {
-    val containerColor = if (primary) PocketShellColors.Accent else PocketShellColors.SurfaceElev
-    val borderColor = if (primary) PocketShellColors.Accent else PocketShellColors.Border
-    val contentColor = if (primary) PocketShellColors.OnAccent else PocketShellColors.Text
-    val alpha = if (enabled) 1f else 0.5f
-    Box(
-        modifier = modifier
-            .clickable(
-                enabled = enabled,
-                role = androidx.compose.ui.semantics.Role.Button,
-                onClick = onClick,
-            )
-            .background(
-                color = containerColor.copy(alpha = alpha),
-                shape = RoundedCornerShape(6.dp),
-            )
-            .border(
-                width = 1.dp,
-                color = borderColor.copy(alpha = alpha),
-                shape = RoundedCornerShape(6.dp),
-            )
-            .padding(horizontal = 12.dp, vertical = 6.dp),
-    ) {
-        Text(
-            text = label,
-            color = contentColor.copy(alpha = alpha),
-            fontSize = 12.sp,
-            fontWeight = FontWeight.SemiBold,
-        )
-    }
-}
-
-/**
- * Issue #180 helper: headline string for the collapsed queue banner.
- * Singular vs plural reads naturally; exposed `internal` so the unit
- * tests can assert against the exact copy without re-implementing the
- * pluralisation logic.
- */
-internal fun pendingSummaryHeadline(items: List<PendingTranscriptionItem>): String = when (items.size) {
-    0 -> ""
-    1 -> "1 pending transcription"
-    else -> "${items.size} pending transcriptions"
-}
-
-/**
- * Issue #180 helper: secondary line summarising the most recent item's
- * status. Avoids cramming a multi-line description into the collapsed
- * row.
- */
-internal fun pendingSummarySubline(items: List<PendingTranscriptionItem>): String {
-    val first = items.firstOrNull() ?: return ""
-    return when {
-        first.isWaitingForNetwork -> "Waiting for network — tap to view"
-        first.atRetryCap -> "Tap to save or discard"
-        first.lastErrorMessage != null -> "Tap to retry"
-        else -> "Tap to retry"
-    }
-}
-
-internal fun outboundQueueHeadline(items: List<OutboundItem>): String = when (items.size) {
-    0 -> ""
-    1 -> "1 unsent prompt"
-    else -> "${items.size} unsent prompts"
-}
-
-internal fun outboundQueueSubline(items: List<OutboundItem>): String {
-    val first = items.firstOrNull() ?: return ""
-    val preview = first.cleanText.trim().lineSequence().firstOrNull().orEmpty()
-    val status = outboundQueueStateLabel(first)
-    return when {
-        preview.isBlank() -> status
-        else -> "$status — $preview"
-    }
-}
-
-internal fun outboundQueueStateLabel(item: OutboundItem): String = when (item.state) {
-    // Issue #971/#987 (Option A): a queued row is waiting to auto-send on
-    // reconnect — show the SINGLE coherent status, not a bare "Queued" that
-    // reads as a mysterious stuck item next to the connection chrome.
-    OutboundState.Queued -> PromptComposerViewModel.WILL_SEND_WHEN_RECONNECTED_MESSAGE
-    OutboundState.Uploading -> "Uploading attachments"
-    OutboundState.InFlight -> "Sending"
-    OutboundState.Delivered -> "Delivered"
-    OutboundState.Failed -> item.lastError?.takeIf { it.isNotBlank() }?.let { "Failed — $it" } ?: "Failed"
-}
-
-internal fun outboundAttachmentCountLabel(count: Int): String =
-    "$count attachment${if (count == 1) "" else "s"}"
-
-internal fun retryableOutboundQueueItem(items: List<OutboundItem>): OutboundItem? =
-    items
-        .filter { it.state == OutboundState.Queued || it.state == OutboundState.Failed }
-        .minByOrNull { it.createdAtMs }
-
-/**
- * Issue #180 helper: humanise an epoch-millis timestamp into "Just now"
- * / "X seconds/minutes/hours ago" relative to [nowMs]. Exposed `internal`
- * so unit tests can verify the breakpoints without standing up the
- * whole banner.
- */
-internal fun formatRelativeTimestamp(timestampMs: Long, nowMs: Long): String {
-    val delta = (nowMs - timestampMs).coerceAtLeast(0L)
-    return when {
-        delta < 30_000L -> "Just now"
-        delta < 60_000L -> "${delta / 1_000L} seconds ago"
-        delta < 60L * 60_000L -> {
-            val minutes = delta / 60_000L
-            if (minutes == 1L) "1 minute ago" else "$minutes minutes ago"
-        }
-        delta < 24L * 60L * 60_000L -> {
-            val hours = delta / (60L * 60_000L)
-            if (hours == 1L) "1 hour ago" else "$hours hours ago"
-        }
-        else -> {
-            val days = delta / (24L * 60L * 60_000L)
-            if (days == 1L) "1 day ago" else "$days days ago"
-        }
-    }
-}
-
 @Composable
 private fun GhostButton(
     label: String,
@@ -3479,42 +2929,7 @@ private val SLASH_DROPDOWN_MAX_HEIGHT = 196.dp
  * non-empty and that the retry / discard / save affordances dispatch
  * into the ViewModel.
  */
-internal const val COMPOSER_PENDING_BANNER_TAG = "prompt-composer-pending-banner"
-internal const val COMPOSER_PENDING_TOGGLE_TAG = "prompt-composer-pending-toggle"
 internal const val COMPOSER_PENDING_SAVED_BANNER_TAG = "prompt-composer-pending-saved"
-internal const val COMPOSER_OUTBOUND_QUEUE_BANNER_TAG = "prompt-composer-outbound-queue"
-internal const val COMPOSER_OUTBOUND_QUEUE_TOGGLE_TAG = "prompt-composer-outbound-queue-toggle"
-// Issue #1308: batch "Resend all" button, shown in the expanded queue banner
-// only when >= 2 rows are resendable.
-internal const val COMPOSER_OUTBOUND_QUEUE_RESEND_ALL_TAG = "prompt-composer-outbound-queue-resend-all"
-
-/**
- * Issue #688: status text shown on a pending row while its retry round-trip
- * is in flight, so a Retry tap gives immediate visible feedback instead of
- * looking like a no-op.
- */
-internal const val PENDING_RETRYING_MESSAGE = "Retrying…"
-
-internal fun composerPendingItemRowTestTag(id: String): String =
-    "prompt-composer-pending-row:$id"
-
-internal fun composerPendingRetryTestTag(id: String): String =
-    "prompt-composer-pending-retry:$id"
-
-internal fun composerPendingDiscardTestTag(id: String): String =
-    "prompt-composer-pending-discard:$id"
-
-internal fun composerPendingSaveTestTag(id: String): String =
-    "prompt-composer-pending-save:$id"
-
-internal fun composerOutboundQueueItemRowTestTag(id: String): String =
-    "prompt-composer-outbound-queue-row:$id"
-
-internal fun composerOutboundQueueDeleteTestTag(id: String): String =
-    "prompt-composer-outbound-queue-delete:$id"
-
-internal fun composerOutboundQueueRetryTestTag(id: String): String =
-    "prompt-composer-outbound-queue-retry:$id"
 
 // -- Previews -----------------------------------------------------------------
